Estimating Climate Characteristics Using Interpolation
Context
When working with measurements of continuous variables, it is often the case that the best measurements of continuous variables are samples taken at discrete locations.  This is typically the case with variables related to weather and climate.  In the situation of weather data, observations of numerous variables including temperature and precipitation are taken at discrete weather stations.  These discrete observations are then used to generate estimates for locations where observations were not directly collected.  The process of generating these estimates is referred to as geographic interpolation -- the use of existing data observed at known locations to estimate values of a variable at other locations where no observations were made.  Given that interpolation is a common geospatial data analysis method, GIS software typically provides several useful tools for generating surfaces (rasters) of interpolated estimates from discrete data points.  The purpose of this assignment is to explore how interpolation can be used in GIS to estimate climate characteristics for a series of locations.

Learning Objectives
Upon completing this assignment, you will have learned:

How to create interpolated surfaces from point data;
About the differences between IDW and Kriging, two popular GIS interpolation methods;
How interpolation can be used to estimate values of variables at given locations;
How to use zonal statistics sample a raster's values based on vector feature geometry;
About the seasonal climate characteristics of cities and places in the state of California.

Deliverables
You have been hired by a California residential real estate firm to prepare a consumer-friendly report about the characteristics of the state's summer and winter climate, designed to inform potential homebuyers who have little knowledge of the state's climate.  It is suggested that you accomplish this task using interpolated values of normal monthly high temperatures (collected by NOAA) (Links to an external site.) for the months of January and August to estimate which are the warmest and coolest cities in the state of California during the winter and the summer.  

Your report should contain the following elements:

Two interpolated maps (generated using either IDW or kriging -- your choice) of average high temperature throughout the areas of California covered by weather stations -- one for the month of January (typically one of the coldest months) and one for the month of August (usually the warmest month with minimal marine layer influence);
A list or table of the ten warmest and coolest cities in the state in both months, with estimated high temperatures included;
Information about the climate normals in several of the state's major cities (i.e. Los Angeles, San Francisco, San Diego, San Diego, Sacramento, etc.)  Use your discretion and select cities that you think would be interesting to a general audience looking to buy a home in California.
On the maps, several cities should be labeled and the estimated high temperature in the city should be indicated.  You are free to label the extreme (highest/lowest temperatures) or several of the state's major cities; select the approach that you feel would be most interesting to a general audience.
Estimates for individual cities should be derived by using zonal statistics to extract values from the interpolated temperature surfaces using a shapefile of city locations, not derived from individual weather station temperature values.

Given that this assignment is placing you in the scenario of generating a product for public consumption, presentation matters!  You may wish to put some effort into document layout, and you will likely wish to include some text captions and written information to contextualize the maps and the data that you included.

You are also required to include a brief technical addendum to your report.  Using the other interpolation method (i.e. if you used IDW interpolation for the estimates in your report, use kriging -- or vice versa), generate the same maps and calculate the same list of the ten warmest and coolest cities along with estimated temperatures for at least one of the two months (only one month is required for the technical addendum).  Briefly discuss in writing whether the results generated using the other interpolation method differ substantially, and what differences are noted if so.

NOTE:  If you are more familiar with the climate of a different state and wish to prepare the same report for a state other than California, you are welcome to do so instead.

Suggestion for extra practice and/or experimentation:  Consider repeating this same process with precipitation normals for the same months.  These data can be downloaded at the same time as the temperature normals, and precipitation is another climate variable for which interpolation can be very useful.