Bridging the Darién Gap with Cost Path Analysis
Responsibilities:
The Pan-American Highway stretches from Prudhoe Bay, Alaska, one of the northernmost settlements in North America, to Ushuaia, Argentina, which is widely regarded as being not only the southernmost settlement in Argentina, but the southernmost city in the world.  This highway is continuous between these two points save a small (100 mile) gap in the road separating Panama and Colombia.  The gap lies not on the Panama Canal, as many expect, but in a region known as the Darién Gap (Links to an external site.).  The Darién Gap is notorious for being difficult to traverse -- given that the region is largely a densely wooded rainforest that is largely inhospitable to human settlement, it has served as a refuge for activities that are considered undesirable, if not illegal, in the countries whose border straddles the gap.

Although no road yet traverses the gap, your task for this assignment is to identify several options for a least cost path for a road linking southern Panama with northern Colombia.  In Panama, the most logical jumping-off point for the Trans-Darién Highway is the small town of Yaviza.  In Colombia, the most logical termini for the road are the small towns of Apartadó and Chigorodó.  You must calculate one least-cost path from Yaviza, Panama to Apartadó, Colombia and one path from Yaviza to Chigorodó, and both paths must be mapped in a document that presents both options.

While you are free to make most of the decisions about the process and the outcomes for this assignment, following are some requirements:

Your cost distance analysis should primarily be based upon (1) terrain slope and (2) presence of large water features (i.e. coastline features and large river channels).  These data can all be determined from SRTM Void-Filled DEMs downloadable from the USGS EarthExplorer website.  Large water features can be defined very simply as any areas of land lying below sea level; this simple conceptualization might require some expansion for a thorough analysis, but it is perfectly sufficient for this assignment.  You are, of course, welcome to include any other water features as obstacles.
Cost should increase significantly as terrain slope increases, and cost should increase even more significantly where large water features are present.  The relationship between slope and cost is not linear -- make sure that your analysis accounts for this by either increasing cost for higher slopes using raster reclassification (covered in Intermediate GIS) or by using the raster calculator to generate a surface where slope is raised exponentially.
Your submission should include:
A map created with ArcGIS showing the origin and destination points, the calculated least-cost paths, the borders of Panama and Colombia, and the road networks on both sides of the border;
An elevation profile of each least cost path;
At least one 3D view showing terrain elevation and profile(s) of the road in three dimensions;
A brief write-up presenting the results and describing other data layers that might be helpful for inclusion in the second iteration of a cost distance analysis for building a highway to traverse the Darién Gap.
Remember -- this is meant to be an exploratory first step toward identifying plausible road paths through the Darién Gap, so you do not need to consider all possible variables in this analysis; terrain slope and presence of large water bodies is sufficient, though you are welcome to include as many other characteristics as you wish!