Building an Interactive Tool with ArcGIS Model Builder
Your task is to develop and submit a sea level rise scenario generation tool for ArcGIS using the ArcGIS Model Builder.  Specifically, your tool will allow users to generate a polygon data layer of areas at-risk for sea level rise based on a digital elevation model (DEM) raster that the user provides.  Accomplishing this outcome requires the development of a multi-step model with user inputs in the form of both variable values and data layers.  Using the ArcGIS tool interface, the user should be prompted to provide an input raster dataset and the destination/path for an output file, as well as (1) the desired sea level change as an integer (in feet) and (2) any desired radius (in feet) for defining an additional 'area of concern' around the input dataset.

Your model must perform the following specific operations:

Using the graphical interface of a tool:
Prompt the user to specify a raster dataset (with elevation expressed in meters) as input for the tool;
Prompt the user to enter a sea level change value (in feet; can be a positive or negative integer);
Prompt the user to enter an 'area of concern' radius, in feet;
Prompt the user to specify a path for output of the final operation (will be a polygon layer);
Upon execution, the tool should perform the following operations:
Use the raster calculator to convert the DEM's elevation values to feet;
Use the raster calculator to generate a 0/1 raster where values of zero correspond to areas below the new sea level and values of one correspond to areas not likely to be inundated;
Convert this raster layer to a vector layer using the "Raster to Polygon" tool;
Using the "Make Feature Layer" tool, create a subset of the polygon layer containing only the areas likely to be submerged;
Calculate the final inundation estimate by using a buffer tool to generate a buffer around the polygon data layer generated in the previous step, wherein the buffer radius should be the 'area of concern' radius provided by the user.