Research Proposal
The subject and scope of your proposal are entirely up to you, but be sure to pick a topic that is neither overly simple nor excessively complex. The proposal must substantially incorporate at least one of the technical themes from this class:

Interpolation analysis;
Network and/or service area analysis;
Viewshed analysis and advanced raster processing methods;
Map Series and automation methods in ArcGIS;
GPS data collection;
Cost distance analysis;
Remote data services;
Model development with ArcGIS Model Builder.
It is of course not necessary to incorporate all of these technical themes in your project, but given that this is the capstone activity of Advanced GIS, your proposal should demonstrate your understanding of techniques presented in this course as well as Intermediate GIS and Introduction to GIS.  Students who have already completed Cartography and/or GIS Programming may also choose to incorporate techniques from those courses in the final project.

Key Components
Research proposal formats vary among disciplines and funding agencies, but generally follow the structure below. Your proposal must include:

A clear and concise project title (not "GEOG 181B Final" or "Research Proposal"!)
An introduction providing background and context; this section should include an overview of existing literature on the area of study, in order to demonstrate the contribution of your proposed research
The research question or hypothesis you plan to address
A detailed methods section that clearly outlines the GIS techniques, approaches, and data sources that are needed to address the research question
A reference section (include at least 3 relevant sources, 2 of which must be peer-reviewed); all references should be properly cited in the body of your proposal, using the citation style of your choice
At least 1 map; additional figures (tables, graphs, etc.) that support your background or methods are encouraged
Provide map(s) (and figures) at the end of the document, accompanied by a caption, and reference them in the text (for examples, you may wish to look through academic books or articles).
The map should provide support for your research proposal. It could be used to visualize the important datasets that your research will use, or to illustrate the results of preliminary analyses. A detailed study area map might also be sufficient, depending on the project.
Figures should be used to convey information that is not possible (or more difficult) to share in the text of the document, while making effective use of space. Multiple maps and/or figures are welcome, but be sure to avoid redundancy!
If your proposal utilizes map automation, do not submit an entire atlas of maps! Most calls for proposals would not accept this much content. In this case, it is better to share examples or a figure that previews the automation process. 
Other common components that you may wish to include:

A summary of preliminary results or findings
The broader impacts or implications of your research